+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Send an email when a live stream is about to be deleted
 - Generic video LTI url
 - Display complete error message in frontend
+- Generic classroom LTI url
 
 ### Changed